+++ CHANGELOG.md
@@ -126,6 +126,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2497 Fix cache invalidation of all files after app deployment
 - YLP-2498 Fix chat widget minimal height
 - YLP-2502 History of device parameters is not sorted properly
+- YLP-2503 Device settings history is not taking timezone into account
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team